1. 支持单击文件打开 2. 支持书签所在文件图标展示
