+++ src/constants/configuration.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_BOOKMARK_COLOR = '#0e69d8';
