+++ src/constants/env.ts
@@ -1,0 +1,4 @@
+import {env} from 'vscode';
+// ENV
+export const APP_ROOT = env.appRoot;
+export const APP_NAME = env.appName;
