+++ src/constants/extension.ts
@@ -1,0 +1,5 @@
+export const EXTENSION_ID = 'bookmark-manager';
+export const EXTENSION_NAME = 'Bookmark Manager (BM)';
+
+export const VIRTUAL_SCHEMA = 'bookmarkmanager';
+export const EXTENSION_VIEW_ID = 'bookmark-manager';
