+++ src/utils/index.ts
@@ -1,18 +1,24 @@
 import * as vscode from 'vscode';
 import {v4 as uuid} from 'uuid';
+
 export * from './command';
 export * from './logger';
 export * from './resources';
-export function createID() {
+export * from './bookmark';
+export * from './icon';
+
+export interface IDisposable {
+  dispose(): void;
+}
+
+export function generateUUID() {
   return uuid();
 }
 
 export function getRelativePath(p: string) {
   return vscode.workspace.asRelativePath(p);
 }
-export interface IDisposable {
-  dispose(): void;
-}
+
 export function dispose<T extends IDisposable>(disposables: T[] | T) {
   if (Array.isArray(disposables)) {
     disposables.forEach(it => it.dispose());
